+++ cider-mode.el
@@ -44,6 +44,7 @@
     (define-key map (kbd "C-c C-p") 'cider-pprint-eval-last-expression)
     (define-key map (kbd "C-c C-r") 'cider-eval-region)
     (define-key map (kbd "C-c C-n") 'cider-eval-ns-form)
+    (define-key map (kbd "C-c M-:") 'cider-read-and-eval)
     (define-key map (kbd "C-c C-m") 'cider-macroexpand-1)
     (define-key map (kbd "C-c M-m") 'cider-macroexpand-all)
     (define-key map (kbd "C-c M-n") 'cider-repl-set-ns)